applies Nutomic suggestions
